requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-whisper==20250625
openunmix==1.3.0
packaging==25.0
platformdirs==4.4.0
pooch==1.8.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
retrying==1.4.2
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.3
submitit==1.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
treetable==0.2.5
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
